add support for 386 / x86 (32 bit)

+++ lsplatform/linux_time_386.go
@@ -1,4 +1,4 @@
-//go:build (amd64 || arm64) && (linux || openbsd || dragonfly || solaris || android)
+//go:build 386 && (linux || openbsd || dragonfly || solaris || android)
 
 package lsplatform
 
@@ -9,12 +9,12 @@ import (
 
 func (*LocalPlatform) FileCTime(fileInfo FileInfo) *time.Time {
 	stat := fileInfo.Sys().(*syscall.Stat_t)
-	ctime := time.Unix(stat.Ctim.Sec, stat.Ctim.Nsec)
+	ctime := time.Unix(int64(stat.Ctim.Sec), int64(stat.Ctim.Nsec))
 	return &ctime
 }
 
 func (*LocalPlatform) FileATime(fileInfo FileInfo) *time.Time {
 	stat := fileInfo.Sys().(*syscall.Stat_t)
-	atime := time.Unix(stat.Atim.Sec, stat.Atim.Nsec)
+	atime := time.Unix(int64(stat.Atim.Sec), int64(stat.Atim.Nsec))
 	return &atime
 }
